requirements.txt
PyPDF2
openai>=1.0.0
python-dotenv
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
